- Separating User Roles handler, User Account handler into two handlers. - Managing Admin Access token Expiry date.
