+++ Eskon.Core/Features/AccountFeatures/Commands/Command/AddUserAccountCommand.cs
@@ -1,0 +1,8 @@
+﻿using MediatR;
+using Eskon.Core.Response;
+using Eskon.Domian.DTOs.User;
+
+namespace Eskon.Core.Features.AccountFeatures.Commands.Command
+{
+    public record AddUserAccountCommand(UserRegisterDto UserRegisterDto) : IRequest<Response<UserReadDto>>;
+}
