+++ Eskon.Core/Features/AccountFeatures/Commands/Command/SignInUserCommand.cs
@@ -1,10 +1,8 @@
-﻿
-using Eskon.Core.Response;
+﻿using Eskon.Core.Response;
 using Eskon.Domian.DTOs.User;
-using Eskon.Domian.Entities.Identity;
 using MediatR;
 
-namespace Eskon.Core.Features.UserFeatures.Commands.Command
+namespace Eskon.Core.Features.AccountFeatures.Commands.Command
 {
     public record SignInUserCommand(UserSignInDto UserSignInDto) : IRequest<Response<TokenResponseDto>>;
 }
