+++ Eskon.Core/Features/AccountFeatures/Commands/Command/SignOutUserCommand.cs
@@ -1,7 +1,7 @@
 ﻿using Eskon.Core.Response;
 using MediatR;
 
-namespace Eskon.Core.Features.UserFeatures.Commands.Command
+namespace Eskon.Core.Features.AccountFeatures.Commands.Command
 {
     public record SignOutUserCommand(string refreshToken) : IRequest<Response<string>>;
 }
