--- Eskon.Core/Features/UserFeatures/Commands/Command/AddAdminRoleToUserCommand.cs
@@ -1,9 +1,0 @@
-﻿
-using Eskon.Core.Response;
-using Eskon.Domian.DTOs.User;
-using MediatR;
-
-namespace Eskon.Core.Features.UserFeatures.Commands.Command
-{
-    public record AddAdminRoleToUserCommand(Guid UserToBeAdminId) : IRequest<Response<TokenResponseDto>>;
-}
