--- Eskon.Core/Features/UserFeatures/Commands/Command/DeleteAdminRoleFromUserCommand.cs
@@ -1,8 +1,0 @@
-﻿using Eskon.Core.Response;
-using Eskon.Domian.DTOs.User;
-using MediatR;
-
-namespace Eskon.Core.Features.UserFeatures.Commands.Command
-{
-    public record DeleteAdminRoleFromUserCommand(Guid UserToRemoveAdminId) : IRequest<Response<TokenResponseDto>>;
-}
