--- Eskon.Core/Features/UserFeatures/Commands/Command/EditUserCommand.cs
@@ -1,8 +1,0 @@
-﻿using Eskon.Domian.Entities.Identity;
-using MediatR;
-
-
-namespace Eskon.Core.Features.UserFeatures.Commands.Command
-{
-    //public record EditUserCommand(User User) : IRequest<Response<UserEditedDTO>>;
-}
