+++ Eskon.Core/Features/UserFeatures/Queries/Query/GetAllUsersQuery.cs
@@ -4,7 +4,5 @@ using Eskon.Domian.DTOs.User;
 
 namespace Eskon.Core.Features.UserFeatures.Queries.Query
 {
-    public record GetAllUsersQuery : IRequest<Response<List<UserReadDto>>>
-    {
-    }
+    public record GetAllUsersQuery : IRequest<Response<List<UserReadDto>>>;
 }
