+++ Eskon.Core/Features/UserFeatures/Queries/Query/GetUserByIdQuery.cs
@@ -4,7 +4,5 @@ using Eskon.Domian.Entities.Identity;
 
 namespace Eskon.Core.Features.UserFeatures.Queries.Query
 {
-    public record GetUserByIdQuery(Guid id) : IRequest<Response<User>>
-    {
-    }
+    public record GetUserByIdQuery(Guid id) : IRequest<Response<User>>;
 }
