+++ Eskon.Core/Features/UserRolesFeatures/Commands/Command/AddAdminRoleToUserCommand.cs
@@ -1,0 +1,8 @@
+﻿using Eskon.Core.Response;
+using Eskon.Domian.DTOs.User;
+using MediatR;
+
+namespace Eskon.Core.Features.UserRolesFeatures.Commands.Command
+{
+    public record AddAdminRoleToUserCommand(Guid UserToBeAdminId) : IRequest<Response<string>>;
+}
