+++ Eskon.Core/Features/UserRolesFeatures/Commands/Command/AddOwnerRoleToUserCommand.cs
@@ -1,9 +1,8 @@
-﻿
-using Eskon.Core.Response;
+﻿using Eskon.Core.Response;
 using Eskon.Domian.DTOs.User;
 using MediatR;
 
-namespace Eskon.Core.Features.UserFeatures.Commands.Command
+namespace Eskon.Core.Features.UserRolesFeatures.Commands.Command
 {
     public record AddOwnerRoleToUserCommand(Guid UserToBeOwnerId) : IRequest<Response<TokenResponseDto>>;
 }
