+++ Eskon.Core/Features/UserRolesFeatures/Commands/Command/DeleteOwnerRoleFromUserCommand.cs
@@ -2,7 +2,7 @@
 using Eskon.Domian.DTOs.User;
 using MediatR;
 
-namespace Eskon.Core.Features.UserFeatures.Commands.Command
+namespace Eskon.Core.Features.UserRolesFeatures.Commands.Command
 {
     public record DeleteOwnerRoleFromUserCommand(Guid UserToRemoveOwnerId) : IRequest<Response<TokenResponseDto>>;
 }
